refactor: unify 2-variable model base; split jansen_rit/horn into packages (goal-05)
Structure refactor with no public API or behaviour change.

Unify the model base
- Add private brainmass/_base.py::NeuralMassDynamics owning the one
  duplicated integrator-selection branch (exp_euler per-variable
  exp_euler_step vs braintools.quad.ode_<method>_step(self.derivative, ...)).
- Migrate XY_Oscillator (hopf/sl/vdp), FitzHughNagumoStep, and
  MontbrioPazoRoxinStep onto it via self._solve_step(...). Drop vdp's
  byte-identical derivative override.
- This restores self.derivative on the Montbrio non-exp_euler path, so its
  rk2/rk4 cases now run (the 2 strict xfails in integrator_test were removed).

Split the dumping grounds
- jansen_rit.py (991 L) -> jansen_rit/ package
  (step, connectivity, tr, network).
- horn.py (773 L) -> horn/ package
  (step, connections, variants, sequential); drop orphaned zeros() helper.
- Each package __init__ re-exports every former module member, so all
  brainmass.<pkg>.<name> paths (incl. pickle paths and the tutorial's
  brainmass.horn.AdditiveConn) keep importing.

Resolve goal-01 debts
- LeadfieldReadout.normalize: keep L1 as the back-compat default
  (True == 'l1'), add opt-in 'l2', validate via _resolve_normalize.
- Fix JansenRitNetwork multi-layer unit chaining by stripping units between
  layers (single-layer output unchanged).

Tests
- New import_compat_test.py: __all__ resolves, old module paths resolve,
  top-level is package object, Model->Step shim warns, moved classes
  pytree-flatten.
- New characterization_test.py: seeded exp_euler trajectories pinned against
  pre-refactor origin/main goldens; all 7 touched models are bit-identical.

Suite 311 -> 337 passed; coverage 97.85% -> 98.17%.
